+++ examples/simple_listener.rs
@@ -1,14 +1,13 @@
 use twitch_irc::login::StaticLoginCredentials;
-use twitch_irc::ClientConfig;
-use twitch_irc::TCPTransport;
 use twitch_irc::TwitchIRCClient;
+use twitch_irc::{ClientConfig, SecureWSTransport};
 
 #[tokio::main]
 pub async fn main() {
     // default configuration is to join chat as anonymous.
     let config = ClientConfig::default();
     let (mut incoming_messages, client) =
-        TwitchIRCClient::<TCPTransport, StaticLoginCredentials>::new(config);
+        TwitchIRCClient::<SecureWSTransport, StaticLoginCredentials>::new(config);
 
     // first thing you should do: start consuming incoming messages,
     // otherwise they will back up.