+++ src/transport/mod.rs
@@ -1,6 +1,6 @@
 #[cfg(feature = "transport-tcp")]
 pub mod tcp;
-#[cfg(feature = "transport-wss")]
+#[cfg(feature = "transport-ws")]
 pub mod websocket;
 
 use crate::message::{IRCMessage, IRCParseError};